Experimenting with making Observable a reference type
This is an experiment that contains an embryo of an implementation of
RxCpp where users no longer work with shared_ptr of Observables and
from functions. Instead there are free functions to create sources like
Return, Empty and Throw, and Observable contains member functions for
calling operators that operate on the Observable (e.g. Select).

Some notes on stuff that might be of interest:

- I’ve built this with Clang on Mac OS X Mavericks. To see what is
implemented see testbench.cpp, Return.cpp and Select.cpp. As I
mentioned above, only a fraction of RxCpp has been implemented using
this approach.

- The design of the concept are to a large extent taken from RxJava.

- Operators are no longer Observables (they don’t inherit from
Observable) they only provide an ”onSubscribe function” (Producer no
longer inherits Observable)

- The free function to create a Return observable is called just, since
return is a C++ keyword (the name is taken from RxJava).

- The free function to create a Throw observable is called error, since
throw is a C++ keyword (the name is taken from RxJava).

- Observable was extracted to its own file to avoid problems with
inclusion order

- I’ve kept the capital letter in one of the overloads to subscribe in
Observable, as this function is called many times throughout the code.